pyproject.toml
[project]
name = "aidevs4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "asyncio>=4.0.0",
    "fastapi[standard]>=0.135.1",
    "google-cloud-aiplatform>=1.141.0",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "mcp[cli]>=1.26.0",
    "openai>=2.20.0",
    "pandas>=3.0.1",
    "polars>=1.38.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.66.0",
    "pydantic-evals[logfire]>=1.66.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
]
